feat & rename : api_client_token으로 이름 변경 및 입력 페이지 이동 시 인가 추가
